feature: follow Qt closer in terms types used
GSK and Evas are specifically 2D and raster focused, Qt's Scene Graph on
the other hand is very interesting in that it provides a QSGGeometry node
which is given two QSGMaterial properties (textures, colors).

A hard decision needs to be made to either be focused on 2D primatives
(rectangles in GSK and Evas) or true triangle based Geometry (Qt).

This begins going down the road of Geometry and Material types following
Qt's footsteps as I believe it is better suited to more ideas.